requirements.txt
streamlit
pandas
google-generativeai
pypdf
python-dotenv
beautifulsoup4
requests
